update: format number
