+++ src/components/header_bar/index.tsx
@@ -1,11 +1,10 @@
 import React from 'react';
 import { DesktopHeaderBar } from 'big-dipper-internal-ui';
+import { formatDenom } from '@utils';
 import { chainConfig } from '@src/chain_config';
 import { HeaderBarProps } from './types';
 import { useHeaderBarHook } from './hooks';
-import {
-  formatDenom, getBasePool,
-} from './utils';
+import { getBasePool } from './utils';
 
 /**
  * Custom Header component
@@ -14,8 +13,7 @@ const HeaderBar = (props:HeaderBarProps) => {
   const { title } = props;
   const { communityPool } = useHeaderBarHook();
   const baseCommunityPool = getBasePool(chainConfig.base, communityPool.data);
-  const communityPoolValue = formatDenom('daric', baseCommunityPool.amount);
-  console.log(communityPoolValue, 'whats the value');
+  const communityPoolValue = formatDenom('daric', baseCommunityPool?.amount ?? 0);
   return (
     <DesktopHeaderBar
       title={title}
@@ -34,7 +32,7 @@ const HeaderBar = (props:HeaderBarProps) => {
         },
         {
           key: 'Community Pool',
-          value: '3,400,000 ATOM',
+          value: `${communityPoolValue.format} ${communityPoolValue.denom}`,
         },
       ]}
     />
