+++ src/utils/index.ts
@@ -1,9 +1,13 @@
 import getLanguageValue from './get_language_value';
 import { handleSearchbarSubmit } from './handle_searchbar';
 import { useGetScreenSize } from './get_screen_size';
+import { formatDenom } from './format_denom';
+import { formatNumber } from './format_number';
 
 export {
   getLanguageValue,
   handleSearchbarSubmit,
   useGetScreenSize,
+  formatNumber,
+  formatDenom,
 };
